+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ This file summarizes the most important modifications to AwesomeShop...
 +------------+---------+------------------------------------------------------+
 | Date       | Type    | Subject                                              |
 +============+=========+======================================================+
+| 2016-06-24 | Feature | Allow the customer to cancel an order                |
++------------+---------+------------------------------------------------------+
 | 2016-06-21 | Feature | Shipping fee estimation for anonymous users          |
 |            |         +------------------------------------------------------+
 |            |         | Display a popup when products are added to the cart  |
@@ -32,3 +34,8 @@ The following are ideas of features that may be implemented in AwesomeShop.
 * Change how translations work in order to get rid of Flask for the frontend
 * "Edit" button on public pages when the user is an admin
 * Send an email when an order is created
+
+Bugs
+====
+
+* Orders list is not updated when a new order is created